+++ src/MiloneLspServer/LspTests.fs
@@ -989,7 +989,9 @@ let private doTestCompletion (p: LLS.ProjectInfo) (wa: LLS.WorkspaceAnalysis) ti
     match pa
           |> LLS.ProjectAnalysis.completion (WorkspaceAnalysis.getModules p wa) docId targetPos
       with
-    | [], pa -> debugProject pa
+    | [], pa ->
+      let output, pa = debugProject pa
+      "No completion results.\n" + output, pa
 
     | result, pa ->
       // FIXME: test kind of items
@@ -1169,6 +1171,62 @@ let private testCompletion () =
         """ ]
       [ "F1"; "F2"; "Length" ] ]
 
+let private testRecordCompletion () =
+  [ testCompletionSingleFile
+      "recordCompletion: record expression"
+      """
+        type R =
+          { A: int
+            B: int }
+
+        let _ : R =
+          { A = 0
+            //
+        //  ^cursor
+            }
+      """
+      [ "A"; "B" ]
+
+    testCompletionSingleFile
+      "recordCompletion: nested record expression"
+      """
+        type R1 = { A: int; B: int }
+        type R2 = { C: int; D: R1 }
+
+        let _ : R2 =
+          { C = 0
+            D = { A = 0 } }
+        //        ^cursor
+      """
+      [ "A"; "B" ]
+
+    testCompletionSingleFile
+      "recordCompletion: empty record expression"
+      """
+        type R = { F1: int; F2: int }
+
+        let takeR (r: R) = ()
+
+        let () =
+          takeR { }
+        //        ^cursor
+      """
+      [ "F1"; "F2" ]
+
+    testCompletionSingleFile
+      "recordCompletion: navigation from a record"
+      """
+        type R = { N1: int; N2: int }
+
+        let makeR () : R = todo
+
+        let _ =
+          let r = makeR ()
+          r.
+        //  ^cursor
+      """
+      [ "Length"; "N1"; "N2" ] ]
+
 // -----------------------------------------------
 // Find projects
 // -----------------------------------------------
@@ -1286,6 +1344,7 @@ let lspTests host =
       testDocumentSymbol ()
       testCodeAction ()
       testCompletion ()
+      testRecordCompletion ()
       testFindProjects () ]
     |> List.collect id
     |> runTests